+++ src/render/dom/utils/transform.ts
@@ -8,7 +8,7 @@ export const transformAxes = ["", "X", "Y", "Z"]
  * An ordered array of each transformable value. By default, transform values
  * will be sorted to this order.
  */
-const order = ["translate", "scale", "rotate", "skew", "transformPerspective"]
+const order = ["perspective", "translate", "scale", "rotate", "skew"]
 
 /**
  * Used to store the keys of all transforms that will distorted a measured bounding box.
@@ -18,7 +18,7 @@ export const boxDistortingKeys: Set<string> = new Set()
 /**
  * Generate a list of every possible transform key.
  */
-export const transformProps = ["x", "y", "z"]
+export const transformProps = ["transformPerspective", "x", "y", "z"]
 order.forEach(operationKey => {
     const isDistorting = new Set(["rotate", "skew"]).has(operationKey)
     transformAxes.forEach(axesKey => {